+++ __tests__/components/history.spec.ts
@@ -464,7 +464,7 @@ describe('text', () => {
       test('should return transcript', () => {
         const history = createHistoryWithPacket(narracterActionPacket);
 
-        const expected = `User: ${narracterActionPacket.narratedAction.text}`;
+        const expected = `User: *${narracterActionPacket.narratedAction.text}*`;
         const transcript = history.getTranscript();
 
         expect(transcript).toEqual(expected);