+++ src/components/history.ts
@@ -1,7 +1,7 @@
 import { v4 } from 'uuid';
 
 import { DEFAULT_USER_NAME } from '../common/constants';
-import { User } from '../common/data_structures';
+import { Extension, User } from '../common/data_structures';
 import { Character } from '../entities/character.entity';
 import {
   Actor,
@@ -66,12 +66,24 @@ interface EmotionsMap {
   [key: string]: EmotionEvent;
 }
 
+interface InworldHistoryProps<InworldPacketT, HistoryItemT> {
+  extension?: Extension<InworldPacketT, HistoryItemT>;
+}
+
 export class InworldHistory<
   InworldPacketT extends InworldPacket = InworldPacket,
+  HistoryItemT extends HistoryItem = HistoryItem,
 > {
   private history: HistoryItem[] = [];
   private queue: HistoryItem[] = [];
   private emotions: EmotionsMap = {};
+  private extension: Extension<InworldPacketT, HistoryItemT>;
+
+  constructor(props?: InworldHistoryProps<InworldPacketT, HistoryItemT>) {
+    if (props?.extension) {
+      this.extension = props.extension;
+    }
+  }
 
   addOrUpdate({
     characters,
@@ -134,10 +146,12 @@ export class InworldHistory<
         return item.id === chatItem?.id;
       });
 
-      if (currentHistoryIndex >= 0 && chatItem) {
-        this.history[currentHistoryIndex] = chatItem;
+      const item = this.extension?.historyItem?.(packet, chatItem) || chatItem;
+
+      if (currentHistoryIndex >= 0) {
+        this.history[currentHistoryIndex] = item;
       } else {
-        this.history = [...this.history, chatItem!];
+        this.history = [...this.history, item!];
       }
     }
 
@@ -265,10 +279,14 @@ export class InworldHistory<
             this.emotions[item.interactionId]?.behavior?.code || '';
           const emotion = emotionCode ? `(${emotionCode}) ` : '';
 
+          const text =
+            item.type === CHAT_HISTORY_TYPE.NARRATED_ACTION
+              ? `*${item.text}*`
+              : item.text;
           transcript +=
             characterLastSpeaking && isCharacter
               ? item.text
-              : `${prefix}${givenName}: ${emotion}${item.text}`;
+              : `${prefix}${givenName}: ${emotion}${text}`;
           characterLastSpeaking = isCharacter;
           break;
         case CHAT_HISTORY_TYPE.TRIGGER_EVENT: