+++ CHANGELOG.md
@@ -8,11 +8,16 @@
 
 - Remove excess extended capabilities generic type
 
+### Fixed
+
+- Generate proto packet just once to dispatching it to the WebSocket connection
+
 ## [1.5.3] - 2023-08-30
 
 ### Fixed
 
 - Ensure `beforeLoadScene` and `afterLoadScene` are present before call them
+
 ## [1.5.2] - 2023-08-29
 
 ### Added